+++ src/main/java/com/buuz135/simpleclaims/commands/SimpleClaimsPartyCommand.java
@@ -28,7 +28,7 @@ public class SimpleClaimsPartyCommand extends AbstractAsyncCommand {
 
     public SimpleClaimsPartyCommand() {
         super("simpleclaimsparty", "Simple Claims Party Commands" );
-        this.addAliases("scp", "sc-party");
+        this.addAliases("scp", "sc-party", "party");
 
         this.addSubCommand(new CreatePartyCommand());
         this.addSubCommand(new PartyInviteCommand());