+++ src/main/java/com/buuz135/simpleclaims/interactions/ClaimPlaceBucketInteraction.java
@@ -1,0 +1,66 @@
+package com.buuz135.simpleclaims.interactions;
+
+import com.buuz135.simpleclaims.claim.ClaimManager;
+import com.buuz135.simpleclaims.claim.party.PartyInfo;
+import com.hypixel.hytale.codec.builder.BuilderCodec;
+import com.hypixel.hytale.component.CommandBuffer;
+import com.hypixel.hytale.component.Ref;
+import com.hypixel.hytale.component.Store;
+import com.hypixel.hytale.math.vector.Vector3i;
+import com.hypixel.hytale.protocol.InteractionState;
+import com.hypixel.hytale.protocol.InteractionType;
+import com.hypixel.hytale.server.core.entity.InteractionContext;
+import com.hypixel.hytale.server.core.entity.InteractionManager;
+import com.hypixel.hytale.server.core.entity.entities.Player;
+import com.hypixel.hytale.server.core.inventory.ItemStack;
+import com.hypixel.hytale.server.core.modules.interaction.interaction.CooldownHandler;
+import com.hypixel.hytale.server.core.modules.interaction.interaction.config.client.PlaceFluidInteraction;
+import com.hypixel.hytale.server.core.universe.PlayerRef;
+import com.hypixel.hytale.server.core.universe.world.World;
+import com.hypixel.hytale.server.core.universe.world.storage.EntityStore;
+import org.checkerframework.checker.nullness.compatqual.NonNullDecl;
+import org.checkerframework.checker.nullness.compatqual.NullableDecl;
+
+import java.util.function.Predicate;
+
+public class ClaimPlaceBucketInteraction extends PlaceFluidInteraction {
+
+    public static final BuilderCodec<ClaimPlaceBucketInteraction> CUSTOM_CODEC = BuilderCodec.builder(ClaimPlaceBucketInteraction.class, ClaimPlaceBucketInteraction::new, PlaceFluidInteraction.CODEC).build();
+
+    @Override
+    protected void interactWithBlock(@NonNullDecl World world, @NonNullDecl CommandBuffer<EntityStore> commandBuffer, @NonNullDecl InteractionType type, @NonNullDecl InteractionContext context, @NullableDecl ItemStack heldItemStack, @NonNullDecl Vector3i targetBlock, @NonNullDecl CooldownHandler cooldownHandler) {
+        Ref<EntityStore> ref = context.getEntity();
+        Store<EntityStore> store = ref.getStore();
+        Player player = store.getComponent(ref, Player.getComponentType());
+        PlayerRef playerRef = store.getComponent(ref, PlayerRef.getComponentType());
+        Predicate<PartyInfo> defaultInteract = PartyInfo::isBlockInteractEnabled;
+        if (playerRef != null && ClaimManager.getInstance().isAllowedToInteract(playerRef.getUuid(), player.getWorld().getName(), targetBlock.getX(), targetBlock.getZ(), defaultInteract)) {
+            super.interactWithBlock(world, commandBuffer, type, context, heldItemStack, targetBlock, cooldownHandler);
+        } else {
+            context.getState().state = InteractionState.Failed;
+            InteractionManager manager = context.getInteractionManager();
+            if (manager != null && context.getChain() != null) {
+                manager.cancelChains(context.getChain());
+            }
+        }
+    }
+
+    @Override
+    protected void simulateInteractWithBlock(@NonNullDecl InteractionType type, @NonNullDecl InteractionContext context, @NullableDecl ItemStack itemInHand, @NonNullDecl World world, @NonNullDecl Vector3i targetBlock) {
+        Ref<EntityStore> ref = context.getEntity();
+        Store<EntityStore> store = ref.getStore();
+        Player player = store.getComponent(ref, Player.getComponentType());
+        PlayerRef playerRef = store.getComponent(ref, PlayerRef.getComponentType());
+        Predicate<PartyInfo> defaultInteract = PartyInfo::isBlockInteractEnabled;
+        if (playerRef != null && ClaimManager.getInstance().isAllowedToInteract(playerRef.getUuid(), player.getWorld().getName(), targetBlock.getX(), targetBlock.getZ(), defaultInteract)) {
+            super.simulateInteractWithBlock(type, context, itemInHand, world, targetBlock);
+        } else {
+            context.getState().state = InteractionState.Failed;
+            InteractionManager manager = context.getInteractionManager();
+            if (manager != null && context.getChain() != null) {
+                manager.cancelChains(context.getChain());
+            }
+        }
+    }
+
+}